Refactor: Move enum codes to CustomEnum file
