+++ VendingMachineApp/VendingMachineApp/PieGraphView.swift
@@ -49,41 +49,6 @@ class PieGraphView: UIView {
     private lazy var myFontSize: CGFloat = 23.0
     private lazy var myTextRect = CGSize(width: 130, height: 40)
 
-    enum DrawType {
-        case defaultGraph
-        case blackCircle
-        case redrawGraph
-    }
-
-    enum PieColors {
-        case red
-        case green
-        case blue
-        case purple
-
-        static let allValues = [red, green, blue, purple]
-
-        var colorList: [UIColor] {
-            switch self {
-            case .red:
-                return self.makeList(themeColor: UIColor.red)
-            case .green:
-                return self.makeList(themeColor: UIColor.green)
-            case .blue:
-                return self.makeList(themeColor: UIColor.blue)
-            case .purple:
-                return self.makeList(themeColor: UIColor.purple)
-            }
-        }
-
-        func makeList(themeColor: UIColor) -> [UIColor] {
-            var colors: [UIColor] = []
-            for i in 0..<6 {
-                colors.append(themeColor.withAlphaComponent(1.0 - (0.15 * CGFloat(i))))
-            }
-            return colors
-        }
-    }
 
     // MARK: Draw and Make the graph
